requirements.txt
streamlit
python-dotenv
PyPDF2
docx2txt
google-generativeai